Bug 816448 - [App Install] Handle errors during app download & install r=etienne
Add different error messages for different errors. The messages are still
generic but all the keys are present in locale so this may change if necessary
without a change in the code.

Also, the actual error name is displayed in logcat at the INFO level, so that a
developer could know what really happens.