Bug 907757 - Give better error message for missing phone and email
When we pick a contact, we cannot filter whether they have an email or a
phone number, depending on the pick request. This might make the user
pick contacts that are invalid in this case. So we notify this with an
alert. This commit ensure that we bring better error messages in this
case.